Building LLM Applications/Crafting an AI-Powered HR Assistant/requirements.txt
openai>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
pypdf>=4.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
ipykernel>=6.0.0
